% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_genes}
\alias{example_genes}
\title{Example Gene List for H.sapiens}
\format{
List
}
\source{
generated for the example
}
\usage{
example_genes
}
\description{
List of genes.
}
\keyword{datasets}
