% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_matrix}
\alias{example_matrix}
\title{Example Expression Matrix for H. sapiens}
\format{
Matrix with rows representing genes and columns representing samples.
}
\source{
GEO - GSE54152
}
\usage{
example_matrix
}
\description{
Expression matrix with normalized values.
}
\keyword{datasets}
