% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_matrix_Mm}
\alias{example_matrix_Mm}
\title{Example Expression Matrix for M. musculus}
\format{
Matrix with rows representing genes and columns representing samples.
}
\source{
GEO - GSE126629 and GSE132214
}
\usage{
example_matrix_Mm
}
\description{
Expression matrix with normalized values.
}
\keyword{datasets}
