% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_network}
\alias{example_network}
\title{Example Network for H. sapiens}
\format{
Network
}
\source{
generated by ARACNe
}
\usage{
example_network
}
\description{
Network computed from the example expression matrix of H. sapiens.
}
\keyword{datasets}
