% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getInteractorsList}
\alias{getInteractorsList}
\title{Get a list of genes with their interactors}
\usage{
getInteractorsList(dataf, reg_list)
}
\arguments{
\item{dataf}{data frame obtained by the network}

\item{reg_list}{list with the regulator genes}
}
\value{
list
}
\description{
This function extracts gets the list of regulators, and it adds the genes
for which we want to compute the activity with VIPER.
Then it filters the data frame obtained by the network to consider only the
regulators of the gene of interest and, for each regulator, it gets a list of
their interactors and the corresponding scores.
}
\examples{

# Load the list of regulators
data(regulators_list)

# Load the example data frame derived from the network
data(network_df)

# Get the list of the regulators with their interactors
interactors <- getInteractorsList(network_df, regulators_list)

}
