% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{IFNSignature}
\alias{IFNSignature}
\title{Compute ADAR activity with the interferon-specific signature}
\usage{
IFNSignature(mat, object)
}
\arguments{
\item{mat}{Either a numeric expression matrix with genes as rows and samples
as columns or an ExpressionSet from a Seurat object, SingleCellExperiment or
SpatialExperiment}

\item{object}{Character string indicating the type of object, either bulk or
eset}
}
\value{
Matrix with ADAR activity scores
}
\description{
The function returns ADAR activity scores for each sample/spot/cell depending
on the starting expression matrix given as input. It employs a signature
derived from a dataset of interferon-treated HEK293T cell line, which is
ideal to measure ADAR activity in an inflammatory context.
}
\examples{

# Load the regulon
data(interferon_regulon)

# Load the example expression matrix
data(example_matrix)

# Compute ADAR activity
res <- IFNSignature(example_matrix, object = "bulk")

}
