% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{inferenceWithA549Signature}
\alias{inferenceWithA549Signature}
\title{Infer ADAR activity with the A549 cells signature}
\usage{
inferenceWithA549Signature(mat)
}
\arguments{
\item{mat}{expression matrix}
}
\value{
VIPER results
}
\description{
This function is used to infer the protein activity with VIPER using the
signature computed from the A549 cell line.
}
\examples{

# Load the regulon
data(regul_A549_cells)

# Load the example expression matrix
data(vsd_matrix)

# Infer protein activity with VIPER
inferenceWithA549Signature(vsd_matrix)

}
