% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{inferenceWithInterferonSignature}
\alias{inferenceWithInterferonSignature}
\title{Infer ADAR activity with the interferon-specific signature}
\usage{
inferenceWithInterferonSignature(mat)
}
\arguments{
\item{mat}{expression matrix}
}
\value{
VIPER results
}
\description{
This function is used to infer the protein activity with VIPER using the
signature computed from HEK293T cells treated with interferon.
}
\examples{

# Load the regulon
data(regul_interferon_specific)

# Load the example expression matrix
data(vsd_matrix)

# Infer protein activity with VIPER
inferenceWithInterferonSignature(vsd_matrix)

}
