% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{inferenceWithNeuronsSignature}
\alias{inferenceWithNeuronsSignature}
\title{Infer ADAR activity with the neurons-specific signature}
\usage{
inferenceWithNeuronsSignature(mat)
}
\arguments{
\item{mat}{expression matrix}
}
\value{
VIPER results
}
\description{
This function is used to infer the protein activity with VIPER using the
signature computed from hESC h9 neurons.
}
\examples{

# Load the regulon
data(regul_hESC_neurons)

# Load the example expression matrix
data(vsd_matrix)

# Infer protein activity with VIPER
inferenceWithNeuronsSignature(vsd_matrix)

}
