% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{inferenceWithYourSignature}
\alias{inferenceWithYourSignature}
\title{Infer ADAR activity with the dataset-specific signature}
\usage{
inferenceWithYourSignature(adjfile, mat)
}
\arguments{
\item{adjfile}{file .adj with the interactors and their scores}

\item{mat}{expression matrix}
}
\value{
VIPER results
}
\description{
This function is used to infer the protein activity with VIPER using the
signature computed from the original dataset.
}
\examples{

# Load the list of regulators
data(regulators_list)

# Load the example data frame derived from the network
data(network_df)

# Load the example expression matrix
data(vsd_matrix)

# Get the list of the regulators with their interactors
interactors <- getInteractorsList(network_df, regulators_list)

# Get the file in the format requested by VIPER
interactors_adj <- writeAdjFile(interactors, file_name = "interactors")

# Infer protein activity with VIPER
inferenceWithYourSignature(interactors_adj, vsd_matrix)

}
