% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interferon_regulon}
\alias{interferon_regulon}
\title{Regulon computed from HEK293T cells treated with interferon}
\format{
Regulon
}
\source{
computed from the dataset
}
\usage{
interferon_regulon
}
\description{
Object of class "regulon".
}
\keyword{datasets}
