% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_neuronal_regulon}
\alias{mouse_neuronal_regulon}
\title{Regulon computed from M. musculus neurons and whole brain}
\format{
Regulon
}
\source{
computed from the dataset
}
\usage{
mouse_neuronal_regulon
}
\description{
Object of class "regulon".
}
\keyword{datasets}
