% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neuronal_regulon}
\alias{neuronal_regulon}
\title{Regulon computed from hESC h9 line}
\format{
Regulon
}
\source{
computed from the dataset
}
\usage{
neuronal_regulon
}
\description{
Object of class "regulon".
}
\keyword{datasets}
