% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regul_A549_cells}
\alias{regul_A549_cells}
\title{Regulon computed from A549 cells}
\format{
Regulon
}
\source{
computed from the dataset
}
\usage{
regul_A549_cells
}
\description{
Object of class "regulon".
}
\keyword{datasets}
