% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regul_hESC_neurons}
\alias{regul_hESC_neurons}
\title{Regulon computed from hESC h9 line}
\format{
Regulon
}
\source{
computed from the dataset
}
\usage{
regul_hESC_neurons
}
\description{
Object of class "regulon".
}
\keyword{datasets}
