% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regul_interferon_specific}
\alias{regul_interferon_specific}
\title{Regulon computed from HEK293T cells treated with interferon}
\format{
Regulon
}
\source{
computed from the dataset
}
\usage{
regul_interferon_specific
}
\description{
Object of class "regulon".
}
\keyword{datasets}
