% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regulators_list}
\alias{regulators_list}
\title{H. sapiens Regulators List}
\format{
List
}
\source{
Gene Ontology
}
\usage{
regulators_list
}
\description{
List of known regulators.
}
\keyword{datasets}
