% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regulators_list_Mm}
\alias{regulators_list_Mm}
\title{M. musculus Regulators List}
\format{
List
}
\source{
Gene Ontology
}
\usage{
regulators_list_Mm
}
\description{
List of known regulators.
}
\keyword{datasets}
