% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tumoral_regulon}
\alias{tumoral_regulon}
\title{Regulon computed from A549 cells}
\format{
Regulon
}
\source{
computed from the dataset
}
\usage{
tumoral_regulon
}
\description{
Object of class "regulon".
}
\keyword{datasets}
