% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tumoralSignature}
\alias{tumoralSignature}
\title{Compute ADAR activity with the cancer-related signature}
\usage{
tumoralSignature(mat, object)
}
\arguments{
\item{mat}{Either a numeric expression matrix with genes as rows and samples
as columns or an ExpressionSet from a Seurat object, SingleCellExperiment or
SpatialExperiment}

\item{object}{Character string indicating the type of object, either bulk or
eset}
}
\value{
Matrix with ADAR activity scores
}
\description{
The function returns ADAR activity scores for each sample/spot/cell depending
on the starting expression matrix given as input. It employs a signature
derived from the union of two datasets of A549 cell line samples, which is
ideal to measure ADAR activity in a cancer-related context.
}
\examples{

# Load the regulon
data(tumoral_regulon)

# Load the example expression matrix
data(example_matrix)

# Compute ADAR activity
res <- tumoralSignature(example_matrix, object = "bulk")

}
