% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{yourSignature}
\alias{yourSignature}
\title{Infer ADAR activity with the dataset-specific regulon}
\usage{
yourSignature(mat, net, genes, organism = "", file_name = "")
}
\arguments{
\item{mat}{Normalized expression matrix used by ARACNe to compute the
network}

\item{net}{Network obtained from ARACNe}

\item{genes}{List of genes of interest}

\item{organism}{"hsapiens" or "mmusculus"}

\item{file_name}{How to nominate the output file}
}
\value{
Matrix with the activity scores of the genes in the reg_list
}
\description{
This function is used to get ADAR activity scores starting from an expression
matrix with genes as rows and samples as columns, and the gene regulation
network obtained by the expression matrix.
}
\examples{

# Load the example data
data(example_network)
data(example_matrix)
data(example_genes)

# Infer ADAR activity
res <- yourSignature(example_matrix, example_network, example_genes,
organism = "hsapiens", file_name = "interactors_scores")

}
